// frontend/packages/shared/src/hooks/blitz/types/PlayerChampion.ts
export type PlayerChampionList = PlayerChampion[];

export type PlayerChampion = {
  assists: number;
  champion_id: number;
  control_wards_placed: number;
  creep_score_at_lane_end: number;
  creep_score_diff_at_lane_end: number;
  damage_dealt: number;
  damage_healed: number;
  damage_magic_dealt: number;
  damage_physical_dealt: number;
  damage_self_mitigated: number;
  damage_taken: number;
  damage_to_champions: number;
  damage_to_objectives: number;
  damage_to_towers: number;
  damage_true_dealt: number;
  deaths: number;
  double_kills: number;
  first_blood: number;
  first_inhibitor_kill: number;
  first_played: string;
  first_tower_kill: number;
  game_count: number;
  game_duration: number;
  gold_at_lane_end: number;
  gold_diff_at_lane_end: number;
  gold_earned: number;
  gold_spent: number;
  killing_sprees: number;
  kills: number;
  last_played: string;
  lp: number;
  lp_games: number;
  lp_wins: number;
  match_ids: string[];
  minions_killed_enemy_jungle: number;
  minions_killed_neutral: number;
  minions_killed_team_jungle: number;
  minions_killed_total: number;
  penta_kills: number;
  quadra_kills: number;
  queue_id: number;
  riot_season_id: number;
  role: string;
  sight_wards_placed: number;
  time_cc_others: number;
  total_time_cc_dealt: number;
  triple_kills: number;
  vision_score: number;
  wards_killed: number;
  wards_placed: number;
  wards_purchased: number;
  wins: number;
  yellow_trinket_wards_placed: number;
};